conda-requirements.txt
# Use this file to create a conda environment using:
# conda create -n <name> --file conda-requirements.txt

# Mandatory dependencies
tensorflow-gpu
gpflow
iris
tqdm

# build dependencies
setuptools

# testing
unittest

# documentation
sphinx